Move old unionRewAgg to UMapSpec
